+++ src/surrealist/ql/statements/__init__.py
@@ -3,6 +3,7 @@ from .define import DefineEvent, DefineParam, DefineUser
 from .delete import Delete
 from .insert import Insert
 from .live import Live
+from .rebuild_index import RebuildIndex
 from .remove import Remove
 from .select import Select
 from .show import Show
@@ -10,4 +11,4 @@ from .transaction import Transaction
 from .update import Update
 
 __all__ = ("Create", "Delete", "Insert", "Live", "Remove", "Select", "Show", "Update", "Transaction", "DefineParam",
-           "DefineUser", "DefineEvent")
+           "DefineUser", "DefineEvent", "RebuildIndex")